extended to instruction set, added a base class for math instructons, added more test cases for basic instruciton operation.

+++ ExperimentalOS/CPU/CPUKernel/InstructionSet/LoadStoreBase.cs
@@ -33,7 +33,7 @@ namespace LagDaemon.ExperimentalOS.CPU.CPUKernel.InstructionSet
     internal abstract class LoadStoreBase : Instruction
     {
         internal int r1, r2;
-        internal uint address;
+        internal int address;
         internal int value;
 
         /// <summary>
@@ -45,32 +45,32 @@ namespace LagDaemon.ExperimentalOS.CPU.CPUKernel.InstructionSet
         /// <param name="value">An imediate value</param>
         /// <param name="address">An address</param>
         /// <param name="comment">A comment</param>
-        internal LoadStoreBase(InstructionCodes code, int r1, int r2, int value, uint address, string comment)
+        internal LoadStoreBase(InstructionCodes code, int r1, int r2, int value, int address, string comment)
             : base(code, comment)
         {
             this.r1 = r1;
             this.r2 = r2;
             this.address = address;
             this.value = value;
-            if (r2 != Symbols.NullRegister && address != (uint)Symbols.NullAddress)
+            if (r2 != Symbols.NullRegister && address != (int)Symbols.NullAddress)
             {
-                this.Size = 4 + sizeof(uint);
+                this.Size = 4 + sizeof(int);
             }
             else if (r2 != Symbols.NullRegister)
             {
-                this.Size = 3 + sizeof(uint);
+                this.Size = 3 + sizeof(int);
             }
-            else if (address != (uint)Symbols.NullAddress)
+            else if (address != (int)Symbols.NullAddress)
             {
-                this.Size = 3 + sizeof(uint);
+                this.Size = 3 + sizeof(int);
             }
-            else if (r2 == Symbols.NullRegister && address == (uint)Symbols.NullAddress)
+            else if (r2 == Symbols.NullRegister && address == (int)Symbols.NullAddress)
             {
                 this.Size = 3 + sizeof(int);
             }
         }
 
-        internal LoadStoreBase(InstructionCodes code, int r1, int r2, int value, uint address) : this(code, r1, r2, value, address, string.Empty) { }
+        internal LoadStoreBase(InstructionCodes code, int r1, int r2, int value, int address) : this(code, r1, r2, value, address, string.Empty) { }
 
         /// <summary>
         /// When implemented assembles a string into this op code
@@ -90,25 +90,25 @@ namespace LagDaemon.ExperimentalOS.CPU.CPUKernel.InstructionSet
             {
                 if (m1.Success)
                 {
-                    return NewInstruction(factory, int.Parse(m1.Groups["r1"].Value), int.Parse(m1.Groups["r2"].Value), 0, uint.Parse(m1.Groups["address"].Value), m1.Groups["comment"] != null ? m1.Groups["comment"].Value : string.Empty);
+                    return NewInstruction(factory, int.Parse(m1.Groups["r1"].Value), int.Parse(m1.Groups["r2"].Value), 0, int.Parse(m1.Groups["address"].Value), m1.Groups["comment"] != null ? m1.Groups["comment"].Value : string.Empty);
                 }
                 if (m2.Success)
                 {
-                    return NewInstruction(factory, int.Parse(m2.Groups["r1"].Value), int.Parse(m2.Groups["r2"].Value), 0, (uint)Symbols.NullAddress, m2.Groups["comment"] != null ? m2.Groups["comment"].Value : string.Empty);
+                    return NewInstruction(factory, int.Parse(m2.Groups["r1"].Value), int.Parse(m2.Groups["r2"].Value), 0, (int)Symbols.NullAddress, m2.Groups["comment"] != null ? m2.Groups["comment"].Value : string.Empty);
                 }
                 if (m3.Success)
                 {
-                    return NewInstruction(factory, int.Parse(m3.Groups["r1"].Value), Symbols.NullRegister, 0, uint.Parse(m3.Groups["address"].Value), m3.Groups["comment"] != null ? m3.Groups["comment"].Value : string.Empty);
+                    return NewInstruction(factory, int.Parse(m3.Groups["r1"].Value), Symbols.NullRegister, 0, int.Parse(m3.Groups["address"].Value), m3.Groups["comment"] != null ? m3.Groups["comment"].Value : string.Empty);
                 }
                 if (m4.Success)
                 {
-                    return NewInstruction(factory, int.Parse(m4.Groups["r1"].Value), Symbols.NullRegister, int.Parse(m4.Groups["value"].Value), (uint)Symbols.NullAddress, m4.Groups["comment"] != null ? m4.Groups["comment"].Value : string.Empty);
+                    return NewInstruction(factory, int.Parse(m4.Groups["r1"].Value), Symbols.NullRegister, int.Parse(m4.Groups["value"].Value), (int)Symbols.NullAddress, m4.Groups["comment"] != null ? m4.Groups["comment"].Value : string.Empty);
                 }
             }
             throw new InstructionParseException("Incorrect op code for this class: {0}, {1}", this.GetType().FullName, assemblyLine);
         }
 
-        protected abstract Instruction NewInstruction(IInstructionFactory factory, int r1, int r2, int value, uint address, string comment);
+        protected abstract Instruction NewInstruction(IInstructionFactory factory, int r1, int r2, int value, int address, string comment);
 
 
         /// <summary>
@@ -119,7 +119,7 @@ namespace LagDaemon.ExperimentalOS.CPU.CPUKernel.InstructionSet
         {
             StringBuilder builder = new StringBuilder();
             builder.AppendFormat("{0} r{1}", Code, r1);
-            if (r2 != Symbols.NullRegister && address != (ulong)Symbols.NullAddress)
+            if (r2 != Symbols.NullRegister && address != (int)Symbols.NullAddress)
             {
                 builder.AppendFormat(", r{0}, ${1}", r2, address);
             }
@@ -127,11 +127,11 @@ namespace LagDaemon.ExperimentalOS.CPU.CPUKernel.InstructionSet
             {
                 builder.AppendFormat(", r{0}", r2);
             }
-            else if (address != (ulong)Symbols.NullAddress)
+            else if (address != (int)Symbols.NullAddress)
             {
                 builder.AppendFormat(", ${0}", address);
             }
-            else if (r2 == Symbols.NullRegister && address == (ulong)Symbols.NullAddress)
+            else if (r2 == Symbols.NullRegister && address == (int)Symbols.NullAddress)
             {
                 builder.AppendFormat(", {0}", value);
             }
